+++ src/colorizer/constants.ts
@@ -1,35 +1,11 @@
-// TODO: Move to dedicated constants file
 import { Color } from "three";
 
-import {
-  CategoricalFeatureThreshold,
-  DrawMode,
-  FeatureThreshold,
-  NumericFeatureThreshold,
-  PlotRangeType,
-  ScatterPlotConfig,
-  TabType,
-  ThresholdType,
-  VectorConfig,
-  ViewerConfig,
-} from "./types";
+import { DrawMode, PlotRangeType, ScatterPlotConfig, TabType, VectorConfig, ViewerConfig } from "./types";
 
 export const BACKGROUND_COLOR_DEFAULT = 0xffffff;
 export const OUTLINE_COLOR_DEFAULT = 0xff00ff;
 export const OUTLIER_COLOR_DEFAULT = 0xc0c0c0;
 export const OUT_OF_RANGE_COLOR_DEFAULT = 0xdddddd;
-
-export const isDrawMode = (mode: number): mode is DrawMode => {
-  return mode === DrawMode.HIDE || mode === DrawMode.USE_COLOR;
-};
-
-export const isThresholdCategorical = (threshold: FeatureThreshold): threshold is CategoricalFeatureThreshold => {
-  return threshold.type === ThresholdType.CATEGORICAL;
-};
-
-export const isThresholdNumeric = (threshold: FeatureThreshold): threshold is NumericFeatureThreshold => {
-  return threshold.type === ThresholdType.NUMERIC;
-};
 
 export const VECTOR_KEY_MOTION_DELTA = "_motion_";
 
@@ -40,10 +16,6 @@ export const getDefaultVectorConfig = (): VectorConfig => ({
   color: new Color(0x000000),
   scaleFactor: 4,
 });
-
-export const isTabType = (tab: string): tab is TabType => {
-  return Object.values(TabType).includes(tab as TabType);
-};
 
 export const getDefaultViewerConfig = (): ViewerConfig => ({
   showTrackPath: true,
